+++ resources/views/admin/workshop/index.blade.php
@@ -48,8 +48,8 @@
                                 @else
                                     <span class="text-gray-300" title="No pick list template assigned"><i class="fa-solid fa-list-check"></i></span>
                                 @endif
-                                @if($workshop->status === "hidden")
-                                    <a href="#" class="hover:text-primary-color" title="Copy workshop link" x-data x-on:click.prevent="SM.copyToClipboard('{{ route('workshop.show', $workshop) }}')"><i class="fa-solid fa-link"></i></a>
+                                @if((bool) $workshop->is_hidden)
+                                    <a href="#" class="hover:text-primary-color" title="Copy workshop link" x-data x-on:click.prevent="SM.copyToClipboard(@js(route('workshop.show', $workshop)))"><i class="fa-solid fa-link"></i></a>
                                 @endif
                                 <a href="{{ route('admin.workshop.duplicate', $workshop) }}" class="hover:text-primary-color" title="Duplicate"><i class="fa-regular fa-copy"></i></a>
                                 <a href="#" class="hover:text-red-600" x-data x-on:click.prevent="SM.confirmDelete('{{ csrf_token() }}', 'Delete workshop?', 'Are you sure you want to delete this workshop? This action cannot be undone', '{{ route('admin.workshop.destroy', $workshop) }}')" title="Delete"><i class="fa-solid fa-trash"></i></a>